+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog (nionui)
 0.3.21 (UNRELEASED)
 -------------------
 
+- Add text edit widget to declarative.
+
 - Allow window show method to specify position and size.
 
 0.3.20 (2018-12-11)